Move build profiles to own namespace
